pyproject.toml
[tool.poetry]
name = "handlerservice"
version = "0.1.0"
description = ""
authors = ["Andrew <veselov.andrey@rocknblock.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
fastapi-jwt = "^0.3.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
ptpython = "^3.0.26"
pydantic-settings = "^2.2.1"
pyyaml = "^6.0.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
